Move property tests into their own package
